% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGenotypeInternal.R
\name{computeBedCov}
\alias{computeBedCov}
\title{computeBedCov}
\usage{
computeBedCov(myreads, snv)
}
\arguments{
\item{myreads}{a \code{data.frame} of the start and end of reads for only
one chr. Column at least  start and end}

\item{snv}{a \code{data.frame} of snv in only one chr with at
least the column: pos}
}
\value{
a \code{data.frame} of snv in only one chr with a new
column readcount
}
\description{
Coverage at the snv position
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschenes and
Alexander Krasnitz
}
\keyword{internal}
