% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGenotypeInternal.R
\name{genoMatrix}
\alias{genoMatrix}
\title{genoMatrix}
\usage{
genoMatrix(snv, resSim, LAFparent)
}
\arguments{
\item{snv}{a \code{data.frame}}

\item{resSim}{\code{list} output of simulateAllele(...)}
}
\value{
a \code{list} of two  \code{matrix}
matRef: \code{matrix} of the readcount of the reference allele
matAlt \code{matrix} of the readcount of the alt allele
}
\description{
Genere twa matrix matRef and matAlt with the
readcount of the reference allele and readcount
of the allele alt respectivily
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschenes and
Alexander Krasnitz
}
\keyword{internal}
