% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGenotypeInternal.R
\name{parseSegLap}
\alias{parseSegLap}
\title{parseSegLap}
\usage{
parseSegLap(mysegs, snv, seqError = 0.001/3, dProp = NA)
}
\arguments{
\item{mysegs}{a \code{list} of \code{data.frame}, from one chr only
with at least the column start, end, lap. If the segments come from
Facets lap = lcn.em / tcn.em}

\item{snv}{a \code{data.frame}}

\item{seqError}{\code{numeric} lap for the other alleles
for the homozygote Default: 0.001/3}

\item{dProp}{\code{numeric} value of lap out of the segments
Default: \code{NA}}
}
\value{
a \code{data.frame} the snv from the input plus column
lap and seg (the index of the segments)
}
\description{
Add the column lap and seg to the snv dataframe from parseStatMinFreq
Here lap is lower allele proportion for the snv in the sample
(not minimal allele freq in the population)
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschenes and
Alexander Krasnitz
}
\keyword{internal}
