% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSimGenotype.R
\name{parseSelMinFreq}
\alias{parseSelMinFreq}
\title{parseSelMinFreq}
\usage{
parseSelMinFreq(snv, genotype, minFreq = 0.01)
}
\arguments{
\item{snv}{a \code{data.frame} with at least the column
chr, pos, ref, alt, AF, EAS_AF , EUR_AF, AFR_AF, AMR_AF, SAS_AF}

\item{minFreq}{\code{numeric} minima frequency in at least one population}

\item{fileList}{a \code{list} of \code{GRanges}, the segments from multiple
files.}
}
\value{
a \code{list} of snv = \code{data.frame} with
column : chr, pos, ref, alt, gtype
and
listSubset = index of the snv keep from the snv input
}
\description{
Generate a subset of snv
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschenes and
Alexander Krasnitz
}
