% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simAnc.R
\docType{package}
\encoding{UTF-8}
\name{simAnc-package}
\alias{simAnc-package}
\alias{simAnc}
\title{simAnc}
\description{
TODO
}
\seealso{
\itemize{
\item \code{\link{simGenotype}} {function for the simulation of the
genotype}
\item \code{\link{simMixture}} {function to simulate recombination}
}
}
\author{
Pascal Belleau,
Astrid Deschênes and Alexander Krasnitz

Maintainer: Pascal Belleau
}
\keyword{package}
