% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGenotypeInternal.R
\name{simulateAllele}
\alias{simulateAllele}
\title{simulateAllele}
\usage{
simulateAllele(snvCovLap, minCov, nbSim)
}
\arguments{
\item{snvCovLap}{a \code{data.frame} with at least the column:
lap and readcount}

\item{minCov}{a \code{numeric} the min value of readcount to keep the snv}

\item{nbSim}{\code{integer} number of simulation}
}
\value{
a \code{list}
listSNV: the index of the snv keep
}
\description{
Compute the readcount for lower allele proportion
don't need to know which allele it is at this step
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschenes and
Alexander Krasnitz
}
\keyword{internal}
