% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGenotypeChr.R
\name{simuleBasicGenoChr}
\alias{simuleBasicGenoChr}
\title{simuleBasicGenoChr}
\usage{
simuleBasicGenoChr(
  genotype,
  infoSNV,
  mysegs,
  nbSim,
  minCov = 10,
  seqError = 0.001/3,
  dProp = NA
)
}
\arguments{
\item{genotype}{\code{data.frame}}

\item{fileList}{a \code{list} of \code{GRanges}, the segments from multiple
files.}

\item{snv}{a \code{data.frame}}
}
\value{
a \code{}
}
\description{

}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{internal}
