% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSimGenotype.R
\encoding{UTF-8}
\name{simulationGenotypeProfileFacets}
\alias{simulationGenotypeProfileFacets}
\title{simulationGenotypeProfileFacets}
\usage{
simulationGenotypeProfileFacets(
  PATH_OUT,
  PATH_1K,
  patientID,
  fileBed,
  fileFacets,
  filePedSel,
  chr,
  nbSim,
  minCov = 10,
  minFreq = 0.01,
  seqError = 0.001/3,
  dProp = NA
)
}
\arguments{
\item{fileList}{a \code{list} of \code{GRanges}, the segments from multiple
files.}

\item{snv}{a \code{data.frame}}

\item{genotype}{a \code{data.frame}}
}
\value{
a \code{}
}
\description{

}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschênes
}
